Move pack_mat_64 to from common_bitpack.h to bitpack.h, or it causes onnx2bnn to include mat.h
